requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
captum==0.7.0
certifi==2026.2.25
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
exceptiongroup==1.3.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
grad-cam==1.5.5
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079158/work
pandas==2.3.3
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pydicom==3.0.2
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.26
torch==2.11.0
torchaudio==2.11.0
torchvision==0.26.0
tqdm==4.67.3
ttach==0.0.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
